pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.1.1",
    "fortune-python>=1.1.1",
    "garminconnect>=0.2.26",
    "loguru>=0.7.3",
    "ollama>=0.4.8",
    "openai-agents>=0.0.14",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
combine_as_imports = true

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
